hw3p2/requirements.txt
torch==1.7.1+cu110
torchvision==0.8.2+cu110

numpy
scikit-learn
pandas
tqdm

torchsummaryX
python-Levenshtein
wget

kaggle
